chore: disable Nomad restarts
